refactor(context): rewrite runWithBinding and runWithInjection in much more meaningful way
The original code contains some complexity: (1) internal method `enterBinding` violates "the single
concern" principle: it ensures that ResolutionSession exists or created, and also it is conserned
with tracking resolution status. (2) The code failed to express the intent that entering and leaving
binding or injection must happen just before action started and right after it is done. So I managed
to get rid of private method, and without changing public interface managed to rewrite runWith-
methods in sane way.

Signed-off-by: Phoenix The Fallen <thephoenixofthevoid@gmail.com>